backend/requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
